backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn
SQLAlchemy==2.0.32
psycopg2-binary
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
PyJWT==2.9.0
gspread==6.1.2
google-auth==2.33.0
python-pptx==0.6.23
google-api-python-client
google-auth-httplib2
